Refactor previous version setup
