+++ extension-manager-integration-test-java/src/test/java/com/exasol/extensionmanager/itest/PreviousVersionManagerTest.java
@@ -4,6 +4,7 @@ import static java.util.stream.Collectors.toList;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.*;
 import static org.junit.jupiter.api.Assertions.*;
+import static org.mockito.Mockito.verify;
 
 import java.io.IOException;
 import java.io.UncheckedIOException;
@@ -13,6 +14,7 @@ import java.net.http.HttpClient.Redirect;
 import java.nio.file.Files;
 import java.nio.file.Path;
 
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.api.io.TempDir;
@@ -22,40 +24,35 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import com.exasol.exasoltestsetup.ExasolTestSetup;
 
 @ExtendWith(MockitoExtension.class)
-class PreviousVersionTest {
+class PreviousVersionManagerTest {
 
     @Mock
     private ExtensionManagerSetup setupMock;
     @Mock
     private ExasolTestSetup exasolTestSetupMock;
-    @Mock
-    private PreviousVersionManager previousVersionManagerMock;
     @TempDir
     private Path tempDir;
+    private PreviousVersionManager testee;
 
-    PreviousVersion.Builder builder() {
+    @BeforeEach
+    void setup() {
         final HttpClient httpClient = HttpClient.newBuilder().followRedirects(Redirect.NORMAL).build();
-        return new PreviousVersion.Builder(setupMock, exasolTestSetupMock, httpClient, previousVersionManagerMock,
-                tempDir).adapterFileName("adapter").currentVersion("currentVersion").previousVersion("previousVersion")
-                .extensionFileName("extensionFileName").project("project");
+        testee = new PreviousVersionManager(setupMock, exasolTestSetupMock, httpClient, tempDir);
     }
 
     @Test
     void fetchExtension() throws IOException {
-        final PreviousVersion testee = builder().build();
-        testee.fetchExtension(URI.create(
+        final String extensionId = testee.fetchExtension(URI.create(
                 "https://extensions-internal.exasol.com/com.exasol/s3-document-files-virtual-schema/2.6.2/s3-vs-extension.js"));
-        final Path file = tempDir.resolve(testee.getExtensionId());
+        final Path file = tempDir.resolve(extensionId);
         assertAll(() -> assertTrue(Files.exists(file), "file downloaded"),
                 () -> assertThat(Files.size(file), equalTo(20875L)));
-        testee.close();
-        assertFalse(Files.exists(file), "file was deleted during cleanup");
+        verify(setupMock).addFileToCleanupQueue(file);
     }
 
     @Test
     void fetchExtensionFailsForMissingFile() throws IOException {
         final URI uri = URI.create("https://extensions-internal.exasol.com/no-such-file");
-        final PreviousVersion testee = builder().build();
         final IllegalStateException exception = assertThrows(IllegalStateException.class,
                 () -> testee.fetchExtension(uri));
         assertThat(exception.getMessage(),
@@ -66,10 +63,9 @@ class PreviousVersionTest {
     @Test
     void fetchExtensionFails() throws IOException {
         final URI uri = URI.create("https://invalid-url");
-        final PreviousVersion testee = builder().build();
         final UncheckedIOException exception = assertThrows(UncheckedIOException.class,
                 () -> testee.fetchExtension(uri));
-        assertThat(exception.getMessage(), startsWith("E-EMIT-29: Failed to download"));
+        assertThat(exception.getMessage(), startsWith("E-EMIT-31: Failed to download"));
         assertThat(Files.list(tempDir).collect(toList()), empty());
     }
 }
